backend/requirements.txt
flask 
flask-sqlalchemy
openai 
dspy-ai 
yfinance 
pandas 
reportlab 
python-dotenv
langgraph
langchain
langchain-huggingface